+++ bash/decoders/dump1090-mutability.sh
@@ -31,38 +31,10 @@
 #                                                                                   #
 # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # # #
 
-BUILDDIR=${PWD}
+BUILDDIR=$PWD
+DUMP1090DIR="$BUILDDIR/dump1090"
 
-## FUNCTIONS
-
-# Function used to check if a package is install and if not install it.
-ATTEMPT=1
-function CheckPackage(){
-    if (( $ATTEMPT > 5 )); then
-        echo -e "\033[33mSCRIPT HALETED! \033[31m[FAILED TO INSTALL PREREQUISITE PACKAGE]\033[37m"
-        echo ""
-        exit 1
-    fi
-    printf "\e[33mChecking if the package $1 is installed..."
-    if [ $(dpkg-query -W -f='${Status}' $1 2>/dev/null | grep -c "ok installed") -eq 0 ]; then
-        if (( $ATTEMPT > 1 )); then
-            echo -e "\033[31m [PREVIOUS INSTALLATION FAILED]\033[37m"
-            echo -e "\033[33mAttempting to Install the package $1 again in 5 seconds (ATTEMPT $ATTEMPT OF 5)..."
-            sleep 5
-        else
-            echo -e "\033[31m [NOT INSTALLED]\033[37m"
-            echo -e "\033[33mInstalling the package $1..."
-        fi
-        echo -e "\033[37m"
-        ATTEMPT=$((ATTEMPT+1))
-        sudo apt-get install -y $1;
-        echo ""
-        CheckPackage $1
-    else
-        echo -e "\033[32m [OK]\033[37m"
-        ATTEMPT=0
-    fi
-}
+source ../bash/functions.sh
 
 clear
 
@@ -87,25 +59,37 @@ CheckPackage git
 CheckPackage curl
 CheckPackage build-essential
 CheckPackage debhelper
+CheckPackage cron
 CheckPackage rtl-sdr
 CheckPackage librtlsdr-dev
 CheckPackage libusb-1.0-0-dev
 CheckPackage pkg-config
 CheckPackage lighttpd
 
-## DOWNLOAD THE DUMP1090-MUTABILITY SOURCE
+## DOWNLOAD OR UPDATE THE DUMP1090-MUTABILITY SOURCE
 
-echo -e "\033[33m"
-echo "Downloading the source code for dump1090-mutability..."
-echo -e "\033[37m"
-git clone https://github.com/mutability/dump1090.git
+# Check if the git repository already exists locally.
+if [ -d $DUMP1090DIR ] && [ -d $DUMP1090DIR/.git ]; then
+    # A directory with a git repository containing the source code exists.
+    echo -e "\033[33m"
+    echo "Updating the local dump1090-mutability git repository..."
+    echo -e "\033[37m"
+    cd $DUMP1090DIR
+    git pull origin master
+else
+    # A directory containing the source code does not exist in the build directory.
+    echo -e "\033[33m"
+    echo "Cloning the dump1090-mutability git repository locally..."
+    echo -e "\033[37m"
+    git clone https://github.com/mutability/dump1090.git
+fi
 
 ## BUILD THE DUMP1090-MUTABILITY PACKAGE
 
 echo -e "\033[33m"
 echo "Building the dump1090-mutability package..."
 echo -e "\033[37m"
-cd $BUILDDIR/dump1090
+cd $DUMP1090DIR
 dpkg-buildpackage -b
 
 ## INSTALL THE DUMP1090-MUTABILITY PACKAGE
@@ -118,7 +102,7 @@ sudo dpkg -i dump1090-mutability_1.15~dev_*.deb
 
 ## CHECK THAT THE PACKAGE INSTALLED
 
-if [ $(dpkg-query -W -f='${Status}' dump1090-mutability 2>/dev/null | grep -c "ok installed") -eq 0 ]; then
+if [ $(dpkg-query -W -f='${STATUS}' dump1090-mutability 2>/dev/null | grep -c "ok installed") -eq 0 ]; then
     echo "\033[31m"
     echo "The dump1090-mutability package did not install properly!"
     echo -e "\033[33m"
@@ -128,13 +112,6 @@ if [ $(dpkg-query -W -f='${Status}' dump1090-mutability 2>/dev/null | grep -c "o
     exit 1
 fi
 
-## START DUMP1090-MUTABILITY
-
-echo -e "\033[33m"
-echo "Starting dump1090-mutability..."
-echo -e "\033[37m"
-sudo /etc/init.d/dump1090-mutability start
-
 ## CONFIGURE LIGHTTPD
 
 echo -e "\033[33m"
@@ -143,55 +120,89 @@ echo -e "\033[37m"
 sudo lighty-enable-mod dump1090
 sudo /etc/init.d/lighttpd force-reload
 
+## DUMP1090-MUTABILITY POST INSTALLATION CONFIGURATION
+
+# Set latitude and longitude.
+echo -e "\033[31m"
+echo "SET THE LATITUDE AND LONGITUDE OF YOUR FEEDER"
+echo -e "\033[33m"
+echo "In order for some performance graphs to work properly you will need to"
+echo "set the latitude and longitude of your feeder. If you do not know the"
+echo "latitude and longitude of your feeder you can find out this information"
+echo "by using Geocode by Address tool found on my web site."
+echo ""
+echo "  https://www.swiftbyte.com/toolbox/geocode"
+echo -e "\033[37m"
+read -p "Feeder Latitude: " FEEDERLAT
+read -p "Feeder Longitude: " FEEDERLON
+echo ""
+ChangeConfig "LAT" $FEEDERLAT "/etc/default/dump1090-mutability"
+ChangeConfig "LON" $FEEDERLON "/etc/default/dump1090-mutability"
+
+# Set dump190-mutability's BEAST_INPUT_PORT to 30104.
+echo -e "\033[33m"
+echo "Configuring dump1090-mutability to listen for BEAST input on port 30104..."
+echo -e "\033[37m"
+ChangeConfig "BEAST_INPUT_PORT" "30104" "/etc/default/dump1090-mutability"
+
+## HEYWHATSTHAT.COM TERRAIN LIMIT RINGS
+
+# Check if the heywhatsthis.com range position file has already been downloaded.
+if [ ! -f /usr/share/dump1090-mutability/html/upintheair.json ]; then
+    echo -e "\033[33m"
+    echo "Dump1090-mutability is able to display terrain limit rings using data obtained"
+    echo "from the website http://www.heywhatsthat.com. Some work will be required on your"
+    echo "part including visiting http://www.heywhatsthat.com and generating a new"
+    echo "panorama set to your location."
+    echo -e "\033[37m"
+    read -p "Do you wish to add terrain limit rings to the dump1090 map? [Y/n] " ADDTERRAINRINGS
+
+    if [[ ! $ADDTERRAINRINGS =~ ^[Nn]$ ]]; then 
+        echo -e "\033[31m"
+        echo "READ THE FOLLOWING INSTRUCTION CAREFULLY!"
+        echo -e "\033[33m"
+        echo "To set up terrain limit rings you will need to first generate a panorama on the website"
+        echo "heywhatsthat.com. To do so visit the following URL:"
+        echo ""
+        echo "  http://www.heywhatsthat.com"
+        echo ""
+        echo "Once the webpage has loaded click on the tab titled New panorama. Fill out the required"
+        echo "information in the form to the left of the map."
+        echo ""
+        echo "After submitting the form your request will be put into a queue to be generated shortly."
+        echo "You will be informed when the generation of your panorama has been completed."
+        echo ""
+        echo "Once generated visit your newly created panorama. Near the top left of the page you will"
+        echo "see a URL displayed which will point you to your newly created panorama. Within this URL's"
+        echo "query string you will see ?view=XXXXXXXX where XXXXXXXX is the identifier for this panorama."
+        echo "Enter below the letters and numbers making up the view identifier displayed there."
+        echo ""
+        echo "Positions for terrain rings for both 10,000 and 40,000 feet will be downloaded by this"
+        echo "script once the panorama has been generated and you are ready to continue."
+        echo -e "\033[37m"
+        read -p "Your heywhatsthat.com view identifier: " HEYWHATSTHATVIEWID
+        read -e -p "First ring altitude in meters (default 3048 meters or 10000 feet): " -i "3048" HEYWHATSTHATRINGONE
+        read -e -p "Second ring altitude in meters (default 12192 meters or 40000 feet): " -i "12192" HEYWHATSTHATRINGTWO
+
+        # Download the generated panoramas JSON data.
+        echo -e "\033[33m"
+        echo "Downloading JSON data pertaining to the panorama ID you supplied..."
+        echo -e "\033[37m"
+        sudo wget -O /usr/share/dump1090-mutability/html/upintheair.json "http://www.heywhatsthat.com/api/upintheair.json?id=${HEYWHATSTHATVIEWID}&refraction=0.25&alts=$HEYWHATSTHATRINGONE,$HEYWHATSTHATRINGTWO"
+    fi
+else
+    # Heywhatsthis.com upintheair.json file already exists.
+    echo -e "\033[33m"
+    echo "Exisitng heywhatsthat.com position data found."
+    echo -e "Skipping terrain limit ring setup...\033[37m"
+fi
+
 ## START DUMP1090-MUTABILITY
 
 echo -e "\033[33m"
 echo "Startng dump1090-mutability..."
 echo -e "\033[37m"
 sudo /etc/init.d/dump1090-mutability start
-
-## HEYWHATSTHAT.COM TERRAIN LIMIT RINGS
-
-echo -e "\033[33m"
-echo "Dump1090-mutability is able to display terrain limit rings using data obtained."
-echo "from the website http://www.heywhatsthat.com. Some work will be required on your"
-echo "part including visiting http://www.heywhatsthat.com and generating a new"
-echo "panorama set to your location."
-echo -e "\033[37m"
-read -p "Do you wish to add terrain limit rings to the dump1090 map? [Y/n] " ADDTERRAINRINGS
-
-if [[ ! $ADDTERRAINRINGS =~ ^[Nn]$ ]]; then 
-    echo -e "\033[31m"
-    echo "READ THE FOLLOWING INSTRUCTION CAREFULLY!"
-    echo ""
-    echo -e "\033[33m"
-    echo "To set up terrain limit rings you will need to first generate a panorama on the website"
-    echo "heywhatsthat.com. To do so visit the following URL:"
-    echo ""
-    echo "  http://www.heywhatsthat.com"
-    echo ""
-    echo "Once the webpage has loaded click on the tab titled New panorama. Fill out the required"
-    echo "information in the form to the left of the map."
-    echo ""
-    echo "After submitting the form your request will be put into a queue to be generated shortly."
-    echo "You will be informed when the generation of your panorama has been completed."
-    echo ""
-    echo "Once generated visit your newly created panorama. Near the top left of the page you will"
-    echo "see a URL displayed which will point you to your newly created panorama. Within this URL's"
-    echo "query string you will see ?view=XXXXXXXX where XXXXXXXX is the identifier for this panorama."
-    echo "Enter below the letters and numbers making up the view identifier displayed there."
-    echo ""
-    echo "Positions for terrain rings for both 10,000 and 40,000 feet will be downloaded by this"
-    echo "script once the panorama has been generated and you are ready to continue."
-    echo -e "\033[37m"
-    read -p "Your heywhatsthat.com view identifier: " HEYWHATSTHATVIEWID
-
-    # Download the generated panoramas JSON data.
-    echo -e "\033[33m"
-    echo "Downloading JSON data pertaining to the panorama ID you supplied..."
-    echo -e "\033[37m"
-    sudo wget -O /usr/share/dump1090-mutability/html/upintheair.json "http://www.heywhatsthat.com/api/upintheair.json?id=${HEYWHATSTHATVIEWID}&refraction=0.25&alts=3048,12192"
-fi
 
 ## DISPLAY MESSAGE STATING DUMP1090-MUTABILITY SETUP IS COMPLETE